Broke up moderngl.__init__.py into modules. Made setup.py executable. Removed unused imports.
