+++ docs/the_guide/first.4.py
@@ -1,5 +1,6 @@
 import moderngl
 import numpy as np
+
 from PIL import Image
 
 ctx = moderngl.create_standalone_context()
