+++ moderngl/program_members/__init__.py
@@ -1,0 +1,5 @@
+from .attribute import *
+from .subroutine import *
+from .uniform import *
+from .uniform_block import *
+from .varying import *
