+++ moderngl/renderbuffer.py
@@ -1,0 +1,108 @@
+__all__ = ['Renderbuffer']
+
+
+class Renderbuffer:
+    '''
+        Renderbuffer objects are OpenGL objects that contain images.
+        They are created and used specifically with :py:class:`Framebuffer` objects.
+        They are optimized for use as render targets, while :py:class:`Texture` objects
+        may not be, and are the logical choice when you do not need to sample
+        from the produced image. If you need to resample, use Textures instead.
+        Renderbuffer objects also natively accommodate multisampling.
+
+        A Renderbuffer object cannot be instantiated directly, it requires a context.
+        Use :py:meth:`Context.renderbuffer` or :py:meth:`Context.depth_renderbuffer`
+        to create one.
+    '''
+
+    __slots__ = ['mglo', '_size', '_components', '_samples', '_depth', '_dtype', '_glo', 'ctx', 'extra']
+
+    def __init__(self):
+        self.mglo = None
+        self._size = (None, None)
+        self._components = None
+        self._samples = None
+        self._depth = None
+        self._dtype = None
+        self._glo = None
+        self.ctx = None
+        self.extra = None
+        raise TypeError()
+
+    def __repr__(self):
+        return '<Renderbuffer: %d>' % self.glo
+
+    def __eq__(self, other):
+        return type(self) is type(other) and self.mglo is other.mglo
+
+    @property
+    def width(self) -> int:
+        '''
+            int: The width of the renderbuffer.
+        '''
+
+        return self._size[0]
+
+    @property
+    def height(self) -> int:
+        '''
+            int: The height of the renderbuffer.
+        '''
+
+        return self._size[1]
+
+    @property
+    def size(self) -> tuple:
+        '''
+            tuple: The size of the renderbuffer.
+        '''
+
+        return self._size
+
+    @property
+    def samples(self) -> int:
+        '''
+            int: The samples of the renderbuffer.
+        '''
+
+        return self._samples
+
+    @property
+    def components(self) -> int:
+        '''
+            int: The components of the renderbuffer.
+        '''
+
+        return self._components
+
+    @property
+    def depth(self) -> bool:
+        '''
+            bool: Is the renderbuffer a depth renderbuffer?
+        '''
+
+        return self._depth
+
+    @property
+    def dtype(self) -> str:
+        '''
+            str: Data type.
+        '''
+
+        return self._dtype
+
+    @property
+    def glo(self) -> int:
+        '''
+            int: The internal OpenGL object.
+            This values is provided for debug purposes only.
+        '''
+
+        return self._glo
+
+    def release(self) -> None:
+        '''
+            Release the ModernGL object.
+        '''
+
+        self.mglo.release()
