+++ setup.py
@@ -1,8 +1,9 @@
+#!/usr/bin/env python
 import platform
 import re
 import sys
 
-from setuptools import Extension, setup
+from setuptools import Extension, find_packages, setup
 
 # pylint: disable=C0103, W0212
 
@@ -137,8 +138,11 @@ setup(
     license='MIT',
     classifiers=classifiers,
     keywords=keywords,
-    packages=['moderngl'],
+    packages=find_packages(),
     install_requires=install_requires,
     ext_modules=[mgl],
     platforms=['any'],
+    python_requires='>=3.4',
+    setup_requires=['pytest-runner'],
+    tests_require=['pytest'],
 )
