+++ tests/test_documentation.py
@@ -20,10 +20,13 @@ class TestCase(unittest.TestCase):
         self.assertSetEqual(implemented - documented - ignored, set(), msg='Implemented but not Documented')
         self.assertSetEqual(documented - implemented, set(), msg='Documented but not Implemented')
 
+        attribute_access = re.compile(r"\w+\.(\w+)")
         for method, docsig in methods:
             classname, methodname = method.split('.')
             sig = str(inspect.signature(getattr(getattr(moderngl, classname), methodname)))
-            sig = sig.replace('self, ', '').replace('moderngl.', '').replace('typing.', '').replace(' -> None', '')
+            sig = sig.replace('self, ', '').replace(' -> None', '')
+            for _ in range(2):
+                sig = re.sub(attribute_access, r'\1', sig)
             sig = sig.replace('(self)', '()').replace(', *,', ',').replace('(*, ', '(')
             sig = re.sub(r'-> \'(\w+)\'', r'-> \1', sig)
             self.assertEqual(docsig, sig, msg=filename + '::' + method)
